Chained Hashing Approach-1
The idea is to store the elements in the hash table to reduce the space. To avoid a collision, chaining is used.

 

Make a hash table of size 1000.
For the addition of a value to a hash table, find its index by passing it in a hash function. In the hash function, the modulus of the value and a constant is taken. The result is the index where the element will be stored in the hash table.
To avoid a collision, use chaining. In chaining, each cell of the hash table points to a linked list of records that have the same hash function value. So, to add the element, just push the element in the list at the found index.
To remove an element, first find its index in the hash table, then search the element in the list of records present at that index. If the element is found, erase the element and return the removed element. If the element is not present already, return -1.
To search an element, first find its index in the hash table. Now, search for the element in the list of records present at the index. If the element is found, return true. Else return false.
Time Complexity
O(Q^2),  ‘Q’ denotes the number of queries. 

 

The time complexity of add operation is O(1).

The time complexity of finding the element is O(Q).

The time complexity of removing the element is O(Q).

The maximum number of elements in a cell in the hashtable will be ‘Q’ because in the worst case, ‘Q’ add operations will be performed.

Space Complexity
O(HASHTABLE_SIZE), ‘HASHTABLE_SIZE’ denotes the size of the hash table.

 

A hashtable is used to store the elements, hence the size required would be of the order of the size of Hashtable.