List Based Approach
Time Complexity
O(Q^2),  ‘Q’ denotes the number of queries that are performed on the HashSet. 

 

The time complexity of add operation is O(1).

The time complexity of finding the element is O(Q).

The time complexity of removing the element is O(Q).

The maximum size of HashSet will be ‘Q’ because in the worst case ‘Q’ add operations will be performed.

Space Complexity
O(Q), where ‘Q’ is the number of queries.

 

In the worst case, the maximum number of elements in the hashtable will be equal to the number of queries.