Mark the elements Approach-3
The idea is to use a boolean vector of size 10^6 to mark the elements in the HashSet.

 

Initialize a boolean vector of size 10^6(size of the queries) with false. Call it ‘HASH’.
For add operation, take the element as an index in hashset as set it to ‘true’.
To check whether the element is present in the HashSet or not, check ‘HASH[ELEMENT]’. If it is ‘true’, the element is present and returns true. Else, return false.
To remove an element from HashSet, turn ‘HASH[ELEMENT]’ to false and return the removed element. If the element is not present in the HashSet, return -1.
Time Complexity
O(1), as it uses constant time.

 

The time complexity of add operation is O(1).

The time complexity of finding the element is O(1).

The time complexity of removing the element is O(1).

Space Complexity
O(ELEMENT), where ‘ELEMENT’ is the largest value that can be inserted in the HashSet.

 

A boolean vector of size 10^6 is used to mark the presence of an element