Power-Set-VVI
Power Set Using Bit-Masking
 

One of the standard and useful method of finding the power set is through bit-masking.

Consider a number with N-bits in its binary representation, if we consider that the state of ith bit depicts whether the ith array element is included in the current subset or not, then we can uniquely identify one of the subsets (as each number has a different binary representation).

Now we can simply iterate from 1 to 2n-1, each number in this iteration will define a subset uniquely. To generate the subset just check for the bits that are ON in binary representation on the number, and for each ON bit, we will simply include an array element corresponding to its position.

Remember to not include an empty subset in the final answer.

 

The steps are as follows :

Declare a 2-D container ans to store all the subsets
Run a for loop for num from 1 to 2^N -1
Run inner for loop for i from 0 to N-1
If the ith bit in num has value equal to 1 then include ith element of the array in the current subset.
Push each subset generated into ans
Finally, return ans
Time Complexity
O( N * 2^N ), where N is the size of the input array.

 

We iterate through N bits for each of the 2^N-1 numbers. Hence the time complexity is O(N * 2^N) 

Space Complexity
O( N * 2^N ), where N is the size of the input array.

 

Since a total of 2^N-1 subsets are generated, and maximum possible length of each subset is of order N . Hence the space complexity is O(N * 2^N).