Used binary search
You have been given a 2-D array 'MAT' of size M x N where 'M' and 'N' denote the number of rows and columns, respectively. The elements of each row are sorted in non-decreasing order.
Moreover, the first element of a row is greater than the last element of the previous row (if exists).
You are given an integer 'TARGET' and your task is to find if it exists in the given 'MAT' or not.