Solved using union find algo, becoz DFS & BFS is slow.
Union-find Algo
In this approach, we will use the union-find algorithm to find the cycle in the undirected graph. 

The main idea behind using this algorithm is that if we have two subsets that are already connected through an edge and if they get connected by another edge then they form a cycle.

 

The union-find algorithm is an algorithm that performs two useful operations on a disjoint-set data structure:

Find: Determine which subset a particular element is in. This can be used for determining if two elements are in the same subset.

Union: Join two subsets into a single subset.

We will iterate through all edges and whenever we find two vertices with the same parent we return “Yes”.

 

Here is the algorithm:

We initialise two arrays ‘PARENT’ and ‘RANK’ to keep track of the parent and rank of the subsets. Here rank denotes the depth of the tree (subset).
Now we will iterate through all edges of the graph:
Find the parent of both vertices (say ‘PARENT1’ and ‘PARENT2’).
If ‘PARENT1’ == ‘PARENT2’
Return “Yes”. Here, ‘PARENT1’ == ‘PARENT2’ represents that both subsets are initially connected and now we have another edge connecting them. Hence a cycle exists.
Else If ‘PARENT1’ != ‘PARENT2’
Union both subsets into a single set. We are doing this because we have two subsets and an edge connecting them. Now both subsets combine and become a single subset.
Finally, return “No”.
Time Complexity
O(M * logN), where ‘N’ is the number of vertices and ‘M’ is the number of edges in the graph.

 

In the worst case, we are calling ‘FIND_PARENT’ function M times and the maximum height of any subset can be log(N). Hence, the time complexity is O(M * logN).

Space Complexity
O(N), where ‘N’ is the number of vertices in the graph.

 

We are using extra arrays ‘PARENT’ and ‘RANK’ each of size N to keep track of the parent and rank of the subsets. Hence, the space complexity is O(N).