Create Roman-Numeral-To-Integer.java
Sample Output 1:
12
27
60
Explanation For Sample Input 1 :
Test Case 1:

We know that ‘X’ is 10 and we have 2 ‘I’ after it. Therefore the number is 12

Test Case 2: 

We know that ‘X’ is 10 and ‘V’ is 5 and we have 2 ‘I’ after it. Therefore the number is 20+5+2=27

Test Case 3:

We see that ‘L’ is 50 and ‘X’ is 10 so the number is 50 +10 =60
Sample Input 2:
2
XC
MM
Sample Output 2:
90
2000