Used hashmap and queue
Given a binary tree, print its bottom view from left to right. Assume, the left and the right child make a 45-degree angle with the parent.
A binary tree is a tree in which each parent node has at most two children.
A node will be in the bottom-view if it is the bottom-most node at its horizontal distance from the root.