Solved using two different approaches
You are given a non-empty binary tree where each node has a non-negative integer value. Return the maximum possible sum of path between any two leaves of the given tree.
The path is also inclusive of the leaf nodes and the maximum path sum may or may not go through the root of the given tree.
If there is only one leaf node in the tree, then return -1.