Solved using 3 diff approaches.
Initially, we were breaking the large problem into small problems but now, let us look at it differently. The idea is to solve the small problem first and then reach the final answer. Thus we will be using a bottom-up approach now. 

 

We are using the value of GRID[i][j] only for calculating the dp values. We will use our GRID vector as our dp table.