TC:O(n*log(m*10)^d), SC:O(1)
You are given two positive integers N and M. You have to find the Nth root of M i.e. M^(1/N).
Note:
N'th root of an integer M is a real number, which when raised to the power N gives M as a result.

N'th root of the M should be correct up to 6 decimal places.