Solved using two diff approaches.
Given that integers are read from a data stream. Your task is to find the median of the elements read so far.
Median is the middle value in an ordered integer list. If the size of the list is even there is no middle value. So the median is the floor of the average of the two middle values.
For Example :
[2,3,4] - median is 3.
[2,3] - median is floor((2+3)/2) = 2.